Moved the operator objects to separate files
