+++ src/solvers/CMakeLists.txt
@@ -7,6 +7,7 @@
 set(solvers_sources
     base_solver.cpp
     nonlinear_solid_solver.cpp
+    nonlinear_solid_operators.cpp
     thermal_solver.cpp
     elasticity_solver.cpp
     )
@@ -14,6 +15,7 @@ set(solvers_sources
 set(solvers_headers
     base_solver.hpp
     nonlinear_solid_solver.hpp
+    nonlinear_solid_operators.hpp
     thermal_solver.hpp
     elasticity_solver.hpp
     )
